Move from gateways.jpa to jpa
