+++ src/main/java/br/com/fiap/soat/techChallenge/gateways/PedidoRepositoryAdapter.java
@@ -1,7 +1,7 @@
 package br.com.fiap.soat.techChallenge.gateways;
 
-import br.com.fiap.soat.techChallenge.gateways.jpa.entity.PedidoEntity;
-import br.com.fiap.soat.techChallenge.gateways.jpa.PedidoJpaRepository;
+import br.com.fiap.soat.techChallenge.jpa.entity.PedidoEntity;
+import br.com.fiap.soat.techChallenge.jpa.PedidoJpaRepository;
 import br.com.fiap.soat.techChallenge.entities.Pedido;
 import br.com.fiap.soat.techChallenge.interfaces.gateways.PedidoRepositoryPort;
 import jakarta.transaction.Transactional;
