+++ src/main/java/br/com/fiap/soat/techChallenge/jpa/ClienteJpaRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiap.soat.techChallenge.gateways.jpa;
+package br.com.fiap.soat.techChallenge.jpa;
 
-import br.com.fiap.soat.techChallenge.gateways.jpa.entity.ClienteEntity;
+import br.com.fiap.soat.techChallenge.jpa.entity.ClienteEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
