+++ src/main/java/br/com/fiap/soat/techChallenge/jpa/PedidoJpaRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiap.soat.techChallenge.gateways.jpa;
+package br.com.fiap.soat.techChallenge.jpa;
 
-import br.com.fiap.soat.techChallenge.gateways.jpa.entity.PedidoEntity;
+import br.com.fiap.soat.techChallenge.jpa.entity.PedidoEntity;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
