+++ src/main/java/br/com/fiap/soat/techChallenge/jpa/ProdutoJpaRepository.java
@@ -1,5 +1,5 @@
-package br.com.fiap.soat.techChallenge.gateways.jpa;
-import br.com.fiap.soat.techChallenge.gateways.jpa.entity.ProdutoEntity;
+package br.com.fiap.soat.techChallenge.jpa;
+import br.com.fiap.soat.techChallenge.jpa.entity.ProdutoEntity;
 import br.com.fiap.soat.techChallenge.entities.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
