+++ src/main/java/br/com/fiap/soat/techChallenge/jpa/entity/ItemDoPedidoEntity.java
@@ -1,4 +1,4 @@
-package br.com.fiap.soat.techChallenge.gateways.jpa.entity;
+package br.com.fiap.soat.techChallenge.jpa.entity;
 
 import br.com.fiap.soat.techChallenge.entities.ItemDoPedido;
 import jakarta.persistence.*;
